Got creature text and status icons into the new draw system.
It would probably be better to disassemble the creature text draw stuff
further, but this will do for now.

That's the last item to be updated to the new draw system.  I declare
this to be a success.  Draw times for my test fort fell from 90ms to
20ms, and several issues were solved along the way.  The only downside
is that now we are carrying around both block data and draw data, and we
only really need the draw data.  The next step will be to remove as much
information as possible from Blocks, and move those sections of the map
loader into Blocs sprite assembly.