Fix mouse click based cursor actions
Previously they did not properly take into
account the y co-ordinate offset caused by the segment's Z-axis
size. This is due in large part to confusion caused by
the two different conventions used for segment's z-positions.

The global `ssState` variable has a convention that Z is 0 on
the top most layer of the segment, while the `segment->segState`
variable uses the convention that Z is zero for one below the
bottom layer.