fix resize crash
The draw event for fog was pushing a pointer to the global fog bitmap, which would be regenerated, thus changing it address, thus leaving the draw event pointing to a freed object.

This was resolved by creating a Fog draw event that does not hold a reference to the fog bitmap; instead the draw event processor will grab the _current_ fog bitmap at the time.

Also replaced `void *` in `draw_event` with a `std::variant` and did some other cleanup around resize processing