improve/fix flow rendering
The code for rendering dragonfire and fire flows, for some reason, tints the sprite used for fire by the density of the flow, but only on the blue and green channels. since the current image is mostly red, this means that density has little effect on the appearance of fire

however, it's worth noting that dragonfire is now (as of v50) blue, at least in the premium graphic set, which means if anybody updates the dragonfire sprite to reflect v50, they may be in for a surprise

this patch switches the effect tint to be blended on all three channels, which may impact the visual effect

in addition, both fire and dragonfire currently render with the same sprite, which may not be desirable in v50. this is not addressed here